+++ packages/atomic/src/components.d.ts
@@ -13,9 +13,11 @@ import { ResultRenderingFunction } from "./components/search/result-lists/result
 import { Section } from "./components/search/atomic-layout-section/sections";
 import { RecommendationEngine } from "@coveo/headless/recommendation";
 import { i18n } from "i18next";
+import { AtomicStore } from "./components/search/atomic-search-interface/store";
 import { InitializationOptions } from "./components/search/atomic-search-interface/atomic-search-interface";
 import { StandaloneSearchBoxData } from "./utils/local-storage-utils";
 import { InsightEngine } from "@coveo/headless/insight";
+import { InitializationOptions as InitializationOptions1 } from "./components/insight/atomic-insight-interface/atomic-insight-interface";
 export namespace Components {
     interface AtomicAriaLive {
         "updateMessage": (region: string, message: string) => Promise<void>;
@@ -649,9 +651,13 @@ export namespace Components {
          */
         "result": Result | FoldedResult;
         /**
+          * Whether an atomic-result-link inside atomic-result should stop propagation.
+         */
+        "stopPropagation"?: boolean;
+        /**
           * Global state for Atomic.
          */
-        "store"?: ReturnType<typeof createAtomicStore>;
+        "store"?: AtomicStore;
     }
     interface AtomicResultBadge {
         /**
@@ -884,7 +890,7 @@ export namespace Components {
          */
         "numberOfQueries": number;
         /**
-          * Defining this option makes the search box standalone.  This option defines the default URL the user should be redirected to, when a query is submitted. If a query pipeline redirect is triggered, it will redirect to that URL instead (see [query pipeline triggers](https://docs.coveo.com/en/1458)).
+          * Defining this option makes the search box standalone (see [Use a Standalone Search Box](https://docs.coveo.com/en/atomic/latest/usage/ssb/)).  This option defines the default URL the user should be redirected to, when a query is submitted. If a query pipeline redirect is triggered, it will redirect to that URL instead (see [query pipeline triggers](https://docs.coveo.com/en/1458)).
          */
         "redirectionUrl"?: string;
         /**
@@ -1129,6 +1135,10 @@ export namespace Components {
          */
         "engine"?: InsightEngine;
         /**
+          * Executes the first search and logs the interface load event to analytics, after initializing connection to the headless search engine.
+         */
+        "executeFirstSearch": () => Promise<void>;
+        /**
           * The service insight interface i18next instance.
          */
         "i18n": i18n;
@@ -1138,9 +1148,13 @@ export namespace Components {
          */
         "iconAssetsPath": string;
         /**
+          * Initializes the connection with the headless insight engine using options for `accessToken` (required), `organizationId` (required), `renewAccessToken`, and `platformUrl`.
+         */
+        "initialize": (options: InitializationOptions) => Promise<void>;
+        /**
           * Initializes the connection with an already preconfigured headless insight engine.
          */
-        "initializeWithRecommendationEngine": (engine: InsightEngine) => Promise<void>;
+        "initializeWithInsightEngine": (engine: InsightEngine) => Promise<void>;
         /**
           * The service insight interface language.
          */
@@ -2460,9 +2474,13 @@ declare namespace LocalJSX {
          */
         "result": Result | FoldedResult;
         /**
+          * Whether an atomic-result-link inside atomic-result should stop propagation.
+         */
+        "stopPropagation"?: boolean;
+        /**
           * Global state for Atomic.
          */
-        "store"?: ReturnType<typeof createAtomicStore>;
+        "store"?: AtomicStore;
     }
     interface AtomicResultBadge {
         /**
@@ -2681,7 +2699,7 @@ declare namespace LocalJSX {
          */
         "numberOfQueries"?: number;
         /**
-          * Defining this option makes the search box standalone.  This option defines the default URL the user should be redirected to, when a query is submitted. If a query pipeline redirect is triggered, it will redirect to that URL instead (see [query pipeline triggers](https://docs.coveo.com/en/1458)).
+          * Defining this option makes the search box standalone (see [Use a Standalone Search Box](https://docs.coveo.com/en/atomic/latest/usage/ssb/)).  This option defines the default URL the user should be redirected to, when a query is submitted. If a query pipeline redirect is triggered, it will redirect to that URL instead (see [query pipeline triggers](https://docs.coveo.com/en/1458)).
          */
         "redirectionUrl"?: string;
         /**