+++ packages/atomic/src/components/common/interface/bindings.ts
@@ -1,13 +1,34 @@
-import {HTMLStencilElement} from '@stencil/core/internal';
+import type {SearchEngine} from '@coveo/headless';
+import type {RecommendationEngine} from '@coveo/headless/recommendation';
+import type {InsightEngine} from '@coveo/headless/insight';
 import {i18n} from 'i18next';
-import {AnyEngineType} from './interface-common';
+import {AtomicCommonStoreData} from './store';
+import {HTMLStencilElement} from '@stencil/core/internal';
+
+export interface CommonStencilStore<StoreData extends AtomicCommonStoreData> {
+  state: StoreData;
+
+  get: <PropName extends keyof StoreData>(
+    propName: PropName
+  ) => StoreData[PropName];
+
+  set: <PropName extends keyof StoreData>(
+    propName: PropName,
+    value: StoreData[PropName]
+  ) => void;
+
+  onChange: <PropName extends keyof StoreData>(
+    propName: PropName,
+    cb: (newValue: StoreData[PropName]) => void
+  ) => () => void;
+}
 
 /**
  * Bindings passed from an interface to its children components.
  */
 export interface CommonBindings<
   Engine extends AnyEngineType,
-  StoreCreation,
+  Store extends CommonStencilStore<AtomicCommonStoreData>,
   InterfaceElement extends HTMLStencilElement
 > {
   /**
@@ -21,9 +42,17 @@ export interface CommonBindings<
   /**
    * Global state for Atomic
    */
-  store: StoreCreation;
+  store: Store;
   /**
    * A reference to the atomic interface element.
    */
   interfaceElement: InterfaceElement;
 }
+
+export type AnyBindings = CommonBindings<
+  AnyEngineType,
+  CommonStencilStore<AtomicCommonStoreData>,
+  HTMLStencilElement
+>;
+
+export type AnyEngineType = SearchEngine | RecommendationEngine | InsightEngine;