+++ packages/atomic/src/components/common/interface/store.ts
@@ -1,14 +1,27 @@
 import {createStore} from '@stencil/store';
+import {AnyEngineType, CommonStencilStore} from './bindings';
 
 export type AtomicCommonStoreData = {
   loadingFlags: string[];
   iconAssetsPath: string;
 };
 
+export interface AtomicCommonStore<StoreData extends AtomicCommonStoreData>
+  extends CommonStencilStore<StoreData> {
+  setLoadingFlag(flag: string): void;
+  unsetLoadingFlag(loadingFlag: string): void;
+  hasLoadingFlag(loadingFlag: string): boolean;
+  waitUntilAppLoaded(callback: () => void): void;
+  isAppLoaded(): boolean;
+  getUniqueIDFromEngine(engine: AnyEngineType): string;
+}
+
 export function createAtomicCommonStore<
   StoreData extends AtomicCommonStoreData
->(initialStoreData: StoreData) {
-  const stencilStore = createStore(initialStoreData);
+>(initialStoreData: StoreData): AtomicCommonStore<StoreData> {
+  const stencilStore = createStore(
+    initialStoreData
+  ) as CommonStencilStore<StoreData>;
 
   return {
     ...stencilStore,
@@ -39,6 +52,13 @@ export function createAtomicCommonStore<
 
     isAppLoaded() {
       return !stencilStore.get('loadingFlags').length;
+    },
+
+    // eslint-disable-next-line @typescript-eslint/no-unused-vars
+    getUniqueIDFromEngine(_engine: AnyEngineType): string {
+      throw new Error(
+        'getUniqueIDFromEngine not implemented at the common store level.'
+      );
     },
   };
 }