+++ packages/atomic/src/components/search/atomic-search-box/atomic-search-box.tsx
@@ -59,6 +59,9 @@ import {Bindings} from '../atomic-search-interface/atomic-search-interface';
  * @part active-suggestion - The currently active suggestion.
  * @part suggestion-divider - An item in the list that separates groups of suggestions.
  * @part suggestion-with-query - An item in the list that will update the search box query.
+ * @part instant-results-item - An instant result rendered by an `atomic-search-box-instant-results` component.
+ * @part instant-result-show-all-button - The button to show all items for the current instant results search rendered by an `atomic-search-box-instant-results` component.
+
  */
 @Component({
   tag: 'atomic-search-box',
@@ -100,7 +103,8 @@ export class AtomicSearchBox {
   @Prop({reflect: true}) public numberOfQueries = 8;
 
   /**
-   * Defining this option makes the search box standalone.
+   * Defining this option makes the search box standalone (see [Use a
+   * Standalone Search Box](https://docs.coveo.com/en/atomic/latest/usage/ssb/)).
    *
    * This option defines the default URL the user should be redirected to, when a query is submitted.
    * If a query pipeline redirect is triggered, it will redirect to that URL instead
@@ -496,15 +500,23 @@ export class AtomicSearchBox {
         break;
       case 'ArrowUp':
         e.preventDefault();
-        this.focusPreviousValue();
+        if (this.firstValue === this.activeDescendantElement) {
+          this.updateActiveDescendant();
+        } else {
+          this.focusPreviousValue();
+        }
         break;
       case 'ArrowRight':
-        e.preventDefault();
-        this.focusPanel(this.rightPanelRef);
+        if (this.activeDescendant || !this.searchBox.state.value) {
+          e.preventDefault();
+          this.focusPanel(this.rightPanelRef);
+        }
         break;
       case 'ArrowLeft':
-        e.preventDefault();
-        this.focusPanel(this.leftPanelRef);
+        if (this.activeDescendant || !this.searchBox.state.value) {
+          e.preventDefault();
+          this.focusPanel(this.leftPanelRef);
+        }
         break;
     }
   }
@@ -598,8 +610,8 @@ export class AtomicSearchBox {
     return part;
   }
 
-  private onSuggestionClick(item: SearchBoxSuggestionElement) {
-    item.onSelect && item.onSelect();
+  private onSuggestionClick(item: SearchBoxSuggestionElement, e: Event) {
+    item.onSelect && item.onSelect(e);
     item.query && this.clearSuggestions();
   }
   private onSuggestionMouseOver(
@@ -652,8 +664,8 @@ export class AtomicSearchBox {
           isSelected ? 'bg-neutral-light' : ''
         }`}
         onMouseDown={(e) => e.preventDefault()}
-        onClick={() => {
-          this.onSuggestionClick(item);
+        onClick={(e: Event) => {
+          this.onSuggestionClick(item, e);
         }}
         onMouseOver={() => {
           this.onSuggestionMouseOver(item, side, id);