+++ packages/headless/src/api/search/search-api-error-response.ts
@@ -5,6 +5,7 @@ export interface SearchAPIErrorWithStatusCode {
   statusCode: number;
   message: string;
   type: string;
+  ignored?: boolean;
 }
 
 export interface SearchAPIErrorWithExceptionInBody {
@@ -21,9 +22,28 @@ function buildDisconnectedError(
   };
 }
 
-export function buildAPIResponseFromErrorOrThrow(error: Error): {
+function buildIgnoredAbortedError(
+  error: DOMException
+): SearchAPIErrorWithStatusCode {
+  return {
+    statusCode: error.code,
+    type: error.name,
+    message: error.message,
+    ignored: true,
+  };
+}
+
+export function buildAPIResponseFromErrorOrThrow(
+  error: Error | DOMException,
+  disableAbortWarning?: boolean
+): {
   error: SearchAPIErrorWithStatusCode;
 } {
+  if (disableAbortWarning && error.name === 'AbortError') {
+    return {
+      error: buildIgnoredAbortedError(error as DOMException),
+    };
+  }
   if (error instanceof DisconnectedError) {
     return {error: buildDisconnectedError(error)};
   }