+++ packages/headless/src/state/state-sections.ts
@@ -42,7 +42,7 @@ import {ExcerptLengthState} from '../features/excerpt-length/excerpt-length-stat
 import {InstantResultsState} from '../features/instant-results/instant-results-state';
 import {InsightConfigurationState} from '../features/insight-configuration/insight-configuration-state';
 import {InsightInterfaceState} from '../features/insight-interface/insight-interface-state';
-import {InsightCaseContextState} from '../features/insight-search/insight-case-context-state';
+import {CaseContextState} from '../features/case-context/case-context-state';
 
 export interface QuerySection {
   /**
@@ -380,5 +380,5 @@ export interface InsightCaseContextSection {
   /**
    * The insight case context to use with the insight search query.
    */
-  insightCaseContext: InsightCaseContextState;
+  insightCaseContext: CaseContextState;
 }